requirements.txt
pandas
pdfminer.six
os
re
csv
scikit-learn
transformers
torch